Move deployAndGetSystem to tests/utils
